+++ src/com/maddyhome/idea/vim/option/OptionChangeListener.java
@@ -18,16 +18,9 @@
 
 package com.maddyhome.idea.vim.option;
 
-import java.util.EventListener;
-
 /**
  * This interface is used for classes that wish to be notified whenever the value of an option has changed
  */
-public interface OptionChangeListener extends EventListener {
-  /**
-   * The value of the option has changed.
-   *
-   * @param event The change event
-   */
-  void valueChange(OptionChangeEvent event);
+public interface OptionChangeListener<T> {
+  void valueChange(T oldValue, T newValue);
 }