+++ CHANGES.md
@@ -19,6 +19,8 @@ usual beta standards.
 [To Be Released]
 --------------
 
+_Available since 0.54.1 EAP:_
+
 **Features:**
 * Surround and Commentary extensions can be repeated with a dot command ([VIM-1118](https://youtrack.jetbrains.com/issue/VIM-1118))
 * Support XDG settings standard ([VIM-664](https://youtrack.jetbrains.com/issue/VIM-664))
@@ -32,6 +34,7 @@ usual beta standards.
 * [VIM-1325](https://youtrack.jetbrains.com/issue/VIM-1325)
   [VIM-1050](https://youtrack.jetbrains.com/issue/VIM-1050)
   [VIM-1627](https://youtrack.jetbrains.com/issue/VIM-1627)
+  [VIM-1867](https://youtrack.jetbrains.com/issue/VIM-1867)
   Fix bindings for active lookup
 * [VIM-1845](https://youtrack.jetbrains.com/issue/VIM-1845) Show ActionGroup popups
 * [VIM-1424](https://youtrack.jetbrains.com/issue/VIM-1424) CTRL-A doesn't have any restrictions now
@@ -40,7 +43,21 @@ usual beta standards.
 * [VIM-1853](https://youtrack.jetbrains.com/issue/VIM-1853) Fix marks for disposed projects
 * [VIM-1858](https://youtrack.jetbrains.com/issue/VIM-1858) Fix imap for autocomplete
 * [VIM-1362](https://youtrack.jetbrains.com/issue/VIM-1362) Search with confirm doesn't scroll down far enough
+
+_Available since 0.54.2 EAP:_
+
+**Fixes:**
+* [VIM-1875](https://youtrack.jetbrains.com/issue/VIM-1875) Fix `isk` in `~/.ideaivmrc`
+* [VIM-1874](https://youtrack.jetbrains.com/issue/VIM-1874) Fix `set clipboard=unnamed` execution from `~/.ideavimrc`
+* [VIM-1878](https://youtrack.jetbrains.com/issue/VIM-1878) Fix `c` command after extract method action
+* [VIM-1884](https://youtrack.jetbrains.com/issue/VIM-1884) Show quickDoc during popup with `CTRL-J`
+
+_Not yet released_
+
+**Fixes:**
 * [VIM-1284](https://youtrack.jetbrains.com/issue/VIM-1284) Fix mapping of digits
+* Fix handling of counts on both operator and motion, e.g. `3d2w` deletes 6 words, instead of 32
+* Allow mapping of `<C-K>` and `<C-V>`/`<C-Q>`
 
 0.54, 2019-11-20
 --------------